pyproject.toml
[tool.poetry]
name = "flask-temp"
version = "0.1.0"
description = ""
authors = ["zhou_mfk <zhou_mfk@163.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
pymysql = "^1.1.1"
types-flask-migrate = "^4.0.0.20240311"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.2.1"
coverage = "^7.5.3"
ruff = "^0.4.6"
mypy = "^1.10.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
disable_warnings = ['no-data-collected']

[tool.coverage.report]
# Regexes for lines to exclude from consideration
exclude_also = [
    # Don't complain about missing debug-only code:
    "def __repr__",
    "if self\\.debug",

    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",

    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstractmethod",
]

ignore_errors = true

[tool.coverage.html]
directory = "coverage_html_report"


[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true


[tool.ruff.lint]
ignore = ["F401"]
